Faciliates usage in a different resolution from when it was calibrated.
This is the first attempt to implement the possibility of using tslib even
from a graphic mode different from the one that was active when calibration
was performed.

*) Modifies ts_calibrate: now the program appends to /etc/pointercal two
additional values, i.e. x and y of current screen resolution.
*) Adds a ts_option function call to tslib API; right now this can be used to
inform tslib about current screen resolution and rotation.
*) Modifies the "linear" plugin to take into account current screen
resolution, and return values scaled accordingly.

Signed-off-by: Alberto Mardegan <mardy@sourceforge.net>
Signed-off-by: Chris Larson <clarson@kergoth.com>